Move request/response structs to graphql client module
