+++ aquadoggo/src/graphql/client/mod.rs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod root;
+mod types;
 
 pub use root::{ClientRoot, EntryArgs};
+pub use types::{EntryArgsRequest, EntryArgsResponse, PublishEntryRequest, PublishEntryResponse};
