--- aquadoggo/src/rpc/mod.rs
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod request;
-mod response;
-
-pub use request::{EntryArgsRequest, PublishEntryRequest};
-pub use response::{EntryArgsResponse, PublishEntryResponse};
